usm: telemetry: tests: Move Clear method to test only file
The Clear method is used only in tests, therefore moving it to a testutil file
to ensure no inappropriate usage is being done with it
