+++ pkg/network/protocols/telemetry/registry_testutil.go
@@ -1,0 +1,19 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2024-present Datadog, Inc.
+
+//go:build test
+
+package telemetry
+
+// Clear metrics
+func Clear() {
+	globalRegistry.Lock()
+	globalRegistry.metrics = nil
+	globalRegistry.Unlock()
+
+	telemetryDelta.mux.Lock()
+	telemetryDelta.stateByClientID = make(map[string]*clientState)
+	telemetryDelta.mux.Unlock()
+}
